requirements.txt
flask==3.1.0
flask-cors==5.0.0
numpy>=1.26.0
Pillow>=10.0.0
firebase-admin==6.5.0
gunicorn==23.0.0
dlib==19.24.6
face-recognition==1.3.0
opencv-python-headless>=4.8.0